Factors out common functionality into the WebsocketApi base class
