+++ ROADMAP.md
@@ -2,14 +2,14 @@
 
 ## v0.2.0 ##
 
-  * OrderBooks
-  * Better config handling
-
-## v0.3.0 ##
-
   * API redesign
       * Unify DataFeeds and Exchanges into RestApiClient and WebsocketApiClient
         for low level API clients.
       * Create high level Feed or Client classes that use composition for
         accessing the low-level api clients.
       * Better subscription handling
+
+## v0.3.0 ##
+
+  * OrderBooks
+  * Better config handling
